Change search_place() to search_places()
I made the name plural because the function returns a list of places in
the json key "candidates", so you're actually searching place[S] rather
than just a place. Whenever the function is used, it can be decided how
that feature wants to handle picking a place (most likely the first one)
but for this function, a list of places will be returned.