Create find_plus_code() to easily get plus codes
This function basically just wraps the search_places() command
(from find_location.py) with code to automatically search for the place
and return the plus code extracted from the return data. It expects the
same parameters as the original search function and forces the
'plus_code' field into the fields parameter by appending the 'plus_code'
field to the 'field' parameter or creating the 'field' key with the
'plus_code' if it doesn't already exist (essentially forcing it).
Finally the function is called within a return statement, followed by a
long trail of subscriptables on the end of it so that the plus code is
instantly and effortlessly extracted then returned all in the same line.